Separate tasks page from settings
